refactor: Stay 메서드 구현체 State로 이동
