+++ src/main/java/domain/user/state/Ready.java
@@ -1,11 +1,8 @@
 package domain.user.state;
 
 import domain.card.Card;
-import domain.game.Winning;
 
 public class Ready extends State {
-
-    public static final String NOT_TERMINATED_YET = "게임 종료 전입니다.";
 
     public Ready() {
         super();
@@ -21,30 +18,5 @@ public class Ready extends State {
             return new Running(cards);
         }
         return this;
-    }
-
-    @Override
-    public boolean isDrawable() {
-        return true;
-    }
-
-    @Override
-    public State stay() {
-        throw new IllegalStateException("게임 시작 전입니다.");
-    }
-
-    @Override
-    public Winning match(State dealer) {
-        throw new IllegalStateException(NOT_TERMINATED_YET);
-    }
-
-    @Override
-    public boolean isBlackJack() {
-        throw new IllegalStateException(NOT_TERMINATED_YET);
-    }
-
-    @Override
-    public boolean isBust() {
-        throw new IllegalStateException(NOT_TERMINATED_YET);
     }
 }
