+++ src/main/java/domain/user/state/Running.java
@@ -1,13 +1,10 @@
 package domain.user.state;
 
 import domain.card.Card;
-import domain.game.Winning;
 import domain.user.Cards;
 
 public class Running extends State {
-
-    public static final String NOT_TERMINATED_YET = "게임 종료 전입니다.";
-
+    
     protected Running(Cards cards) {
         super(cards);
     }
@@ -25,27 +22,7 @@ public class Running extends State {
     }
 
     @Override
-    public boolean isDrawable() {
-        return true;
-    }
-
-    @Override
     public State stay() {
         return new Stay(cards);
-    }
-
-    @Override
-    public Winning match(State dealer) {
-        throw new IllegalStateException(NOT_TERMINATED_YET);
-    }
-
-    @Override
-    public boolean isBlackJack() {
-        throw new IllegalStateException(NOT_TERMINATED_YET);
-    }
-    
-    @Override
-    public boolean isBust() {
-        throw new IllegalStateException(NOT_TERMINATED_YET);
     }
 }
